llm_rag/requirements.txt
langchain>=0.0.267
chromadb>=0.4.13
openai>=1.0.0
pypdf>=3.15.1
tiktoken>=0.5.1
pandas>=2.0.3
numpy>=1.24.3
streamlit>=1.26.0
markdown>=3.4.4
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2 